go compiler: simplify test output and fix int division

+++ tests/machine/x/go/math_ops.go
@@ -8,6 +8,6 @@ import (
 
 func main() {
 	fmt.Println((6 * 7))
-	fmt.Println((7 / 2))
+	fmt.Println((float64(7) / float64(2)))
 	fmt.Println((7 % 2))
 }
